Moved the remaining customisations for C172 et al into the default aircraft
It’s not necessary to have a completely separate profile for the few things that are different. This also requires my autopilot-status PR which fixes the GPSS detection and the retractable-gear PR which identifies whether the aircraft has retractable gear. With those two already applied, this PR should exactly reproduce the C172 et al behaviour from v1.1.1 of the plugin.