B738: remove separate aircraft profile
This is the last separate aircraft profile now removed. There is one change here that effects other aircraft—the anti-ice. It’s now lit when the anti-ice (or pitot heat) warning would be lit in the cabin. I think that’s better than what it was before for, which was it was lit when pitot heat was on.